Ensure that query parameters are properly encoded
The `Uri` class does not encode the passed query string. By using the `query` option from
guzzle, we can ensure that the parameters are properly encoded.
Without proper encoding, the pixxio endpoint responds with a 400 error when a query parameter
contains invalid characters. This can happen especially in the `options` query parameter